+++ include/concepts/board_state_coordinator.hpp
@@ -2,7 +2,7 @@
 
 #include <concepts>
 #include <gameboard/board_data_structs.hpp>
-#include <evaluator_data_structs.hpp>
+#include <moveselection/evaluator_data_structs.hpp>
 #include <integer_types.hpp>
 #include <memory>
 #include <gameboard/move_data_structs.hpp>
