+++ src/core/move_evaluators/move_evaluators.cpp
@@ -1,7 +1,7 @@
 //! @file move_evaluators.cpp
 //! Linkage point for .hpp and .tpp files where template-heavy concrete MoveEvaluators
 //! are implemented.
-#include <move_evaluators.hpp>
+#include <moveselection/move_evaluators.hpp>
 
 namespace moveselection {
 namespace minimaxutils {
