+++ tests/core/minimax_zobrist_component_test.cpp
@@ -1,7 +1,7 @@
 #include <gameboard/game_board.hpp>
 #include <gtest/gtest.h>
 #include <gameboard/move_data_structs.hpp>
-#include <move_evaluators.hpp>
+#include <moveselection/move_evaluators.hpp>
 #include <piecepoints/piece_position_points.hpp>
 #include <boardstate/zobrist.hpp>
