+++ tests/core/move_evaluator_human_for_concepts/move_evaluator_human_for_concepts_test.cpp
@@ -2,8 +2,8 @@
 #include <gtest/gtest.h>
 #include <iostream>
 #include <gameboard/move_data_structs.hpp>
-#include <move_evaluator_human_for_concepts.hpp>
-#include <move_translator.hpp>
+#include <moveselection/move_evaluator_human_for_concepts.hpp>
+#include <movetranslation/move_translator.hpp>
 #include <sstream>
 
 class InputRetrievalMessagesTest : public ::testing::Test {
