+++ tests/core/move_evaluator_minimax_for_concepts/move_evaluator_minimax_for_concepts_test.cpp
@@ -2,7 +2,7 @@
 #include <concepts/move_evaluator.hpp>
 #include <gameboard/game_board_for_concepts.hpp>
 #include <gtest/gtest.h>
-#include <move_evaluator_minimax_for_concepts.hpp>
+#include <moveselection/move_evaluator_minimax_for_concepts.hpp>
 #include <piecepoints/piece_position_points_for_concepts.hpp>
 #include <type_traits>
 #include <boardstate/zobrist_for_concepts.hpp>
