+++ tests/core/transposition_table_pruner_test.cpp
@@ -1,6 +1,6 @@
 #include <gameboard/game_board.hpp>
 #include <gtest/gtest.h>
-#include <move_evaluators.hpp>
+#include <moveselection/move_evaluators.hpp>
 #include <piecepoints/piece_position_points.hpp>
 #include <utility_functs.hpp>
 #include <boardstate/zobrist.hpp>
